Move module reader tests to be integration
